TC 9221: Cleanup OutdoorPvP code

+++ src/game/OutdoorPvPTF.cpp
@@ -18,6 +18,7 @@
 
 #include "OutdoorPvPTF.h"
 #include "OutdoorPvPMgr.h"
+#include "OutdoorPvP.h"
 #include "WorldPacket.h"
 #include "Player.h"
 #include "ObjectMgr.h"
@@ -220,7 +221,7 @@ bool OutdoorPvPTF::SetupOutdoorPvP()
     first_digit = 0;
 
     // add the zones affected by the pvp buff
-    for (int i = 0; i < OutdoorPvPTFBuffZonesNum; ++i)
+    for (uint8 i = 0; i < OutdoorPvPTFBuffZonesNum; ++i)
         RegisterZone(OutdoorPvPTFBuffZones[i]);
 
     AddCapturePoint(new OPvPCapturePointTF(this,TF_TOWER_NW));
